refactor: moved all assets in the public folder
