+++ frontend/src/components/Tree/File.svelte
@@ -1,6 +1,5 @@
 <script>
-    // @ts-ignore
-    import { addTab, TabConfig } from "/src/components/Header/Tabs/TabStore.js";
+    import { addTab, TabConfig } from "../Header/Tabs/TabStore.js";
     export let name;
     export let path;
     export let relativePath;
@@ -22,7 +21,7 @@
         try {
             const report = await getContent();
             if (report.status === "ERROR") throw new Error(report.message);
-            addTab(new TabConfig(name, path, report.content, false));
+            addTab(new TabConfig(name, path, report.content));
         } catch (e) {
             console.error("Failed to load file", e);
         }
@@ -31,7 +30,7 @@
 
 <button
     on:click={clickFile}
-    style="background-image: url(/src/assets/themes/{theme}/icons/{type}.svg), url(/src/assets/themes/{theme}/icons/file.svg)"
+    style="background-image: url(assets/themes/{theme}/icons/{type}.svg), url(assets/themes/{theme}/icons/file.svg)"
     >{name}</button
 >
